move apiv2 docker composer

+++ devops/local/pull_staging.sh
@@ -13,6 +13,7 @@ if ! [[ -x "$(command -v jq)" ]]; then
   exit 1
 fi
 
+
 SCRIPT_DIR=$( cd -- "$( dirname -- "${BASH_SOURCE[0]}" )" &> /dev/null && pwd )
 PROJECT_DIR=$(cd $SCRIPT_DIR/../../ && pwd)
 
@@ -31,7 +32,18 @@ DUMP_DIRECTORY="$SCRIPT_DIR/dump"
 MONGO_STAGING_URL=$($PROJECT_DIR/api/src/scripts/get_secrets.sh snu-ci | jq -r ".MONGO_URL")
 MONGO_STAGING_NAME=bmwmw2zjc1t8eoc2zfy3
 MONGO_LOCAL_NAME=snu_dev
-MONGO_LOCAL_URL="mongodb://root:password123@localhost:27017/$MONGO_LOCAL_NAME?replicaSet=rs0&authSource=admin&directConnection=true";
+printf '\nDo you want to use mongo replica set (y/N)? '
+read answer
+if [ "$answer" = "${answer#[Yy]}" ] ;then 
+    echo "Using single instance..."
+    MONGO_LOCAL_URL="mongodb://localhost:27017/$MONGO_LOCAL_NAME";
+    USE_REPLICA_SET=0
+else
+    echo "Using multiple instance (replica set) $answer..."
+    MONGO_LOCAL_URL="mongodb://root:password123@localhost:27017/$MONGO_LOCAL_NAME?replicaSet=rs0&authSource=admin&directConnection=true";
+    USE_REPLICA_SET=1
+fi
+
 export ES_ENDPOINT=localhost:9200
 
 echo "= DUMP DIR   : $DUMP_DIRECTORY\n"
@@ -55,7 +67,11 @@ if [ "$answer" = "${answer#[Yy]}" ] ;then
     echo "Aborting..."
     exit 1
 fi
-echo "Pulling database..."
+
+echo ""
+echo "==============================="
+echo "= Pulling database..."
+echo "==============================="
 
 ###
 # Dump staging database in local
@@ -66,15 +82,28 @@ fi
 rm -fr $DUMP_DIRECTORY/*
 mongodump --gzip --out=$DUMP_DIRECTORY $MONGO_STAGING_URL
 
-echo "Restoring database..."
+echo ""
+echo "==============================="
+echo "= Restoring database..."
+echo "==============================="
 ###
 # restore local dump in local database
 ###
 mongorestore --drop --gzip $MONGO_LOCAL_URL $DUMP_DIRECTORY/$MONGO_STAGING_NAME
 
-echo "Request full index rebuild..."
-docker-compose -f docker-compose.replica.yml stop es_datariver
-mongosh $MONGO_LOCAL_URL --eval "db.monstache.drop()"
-../ES_datariver/uninstall.sh
-../ES_datariver/install.sh
-docker-compose -f docker-compose.replica.yml restart es_datariver
+echo ""
+echo "==============================="
+echo "= Request full index rebuild..."
+echo "==============================="
+if [ $USE_REPLICA_SET -eq 1 ];then
+  echo "Restarting ES datariver..."
+  docker-compose -f docker-compose.replica.yml stop es_datariver
+  mongosh $MONGO_LOCAL_URL --eval "db.monstache.drop()"
+  ../ES_datariver/uninstall.sh
+  ../ES_datariver/install.sh
+  docker-compose -f docker-compose.replica.yml restart es_datariver
+else
+  echo "Running old script reindex_es_all_models"
+  cd $PROJECT_DIR/api
+  npx ts-node ./src/scripts/reindex_es_all_models/index.js
+fi
